pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "folium>=0.20.0",
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
    "streamlit-folium>=0.27.0",
]
